chore(daemon): socket.rs -> daemon/server.hs
The Server struct had an unnecessary indirection after removing the
varlink backend.
